Use conditional cascade to check render capability in `notes show`.
Pandoc-output HTML displays most thoroughly in `lynx`, so check if it it
exists on the system first to use for `notes show --render`. If `lynx`
is not found in the system environment, fallback to `w3m` if present. If
`w3m` is not found, fallback to displaying the unrendered file using
`$PAGER`.