Refactor 'help' and 'version' to behave like regular subcommands.
These two functions can each be invoked with a lone option flag
regardless of whether the git repository has been initialized, but calls
to the subcommand functions can be used. In other cases, they can be
used like regular subcommand functions.

This structure is more uniform and consistent than using functions
outside of the subcommand structure, and it makes it easier to avoid
repeating the subcommand matching string in `_notes()`.