Add scoping.
When the first argument includes a semicolon, the string before the
semicolon is set as the "scope," which maps directly to a "repo," and
everything after the semicolon is set as the subcommand, which is then
run within the "scope" repository.

Scoping makes it possible to run commands in repositories other than the
current one, without switching repositories with `notes repo use <name>`.