Don't use pager with `help <subcommand>`.
Previously, when the `help <subcommand>` output is longer than the
number of lines in the terminal, the content is displayed in the pager.
This leads to inconsistent behavior, with some help views printed in the
pager while others are printed directly to the terminal.

This change removed the pager logic for `help <subcommand>` so the
output is always printed directly to the terminal.

The pager is still used in `help` without <subcommand> since it is very
long. This should be reviewed in the future.