Skip unindexed folders in `search`.
In some cases, such as an imported folder with subdirectories, folders
with search hits are not indexed, meaning they do not contain a .index
file listing the contents of the directory.

`nb` currently relies on .index files to identify items. When a .index file
is missing, some commands, such as `index`, will automatically create one.
Since `search` calls `index`, it would generate .index files for all
subfolders with results.

To avoid this behavior for now, skip displaying search results for any
folders that do not contain a .index file.